api/python_worker/requirements.txt

# MongoDB client
pymongo>=4.0

# Vector DB
# Pin to the confirmed stable release
chromadb==1.3.2


# OpenAI client - allow the 1.x stable series (keep <2.0 for safety)
openai>=1.0.0,<2.0
# Dimensionality reduction
umap-learn>=0.5

# Numeric / ML helpers
numpy>=1.23
scikit-learn>=1.1

# Utilities
tqdm>=4.60
python-dotenv>=1.0
pyyaml>=6.0
gunicorn>=20.1
fastapi>=0.95
uvicorn>=0.22



